docs(vscode-settings): Adding VSCode configuration filesMerge branch 'feature-release' into refactor/oci-tmp-refactor-cindyguo

+++ processMake.sh
@@ -6,10 +6,14 @@ source $(pwd)/Shells/installVendor.sh && \
 dependenciesTraversal $(pwd)/Settings/.Json/globalDependencies.json $(pwd)/$Vendors/.$Vendors.json
 
 # Cmake process
-rm -rf build
+if [[ "$1" = "--remove" ]]; then
+	rm -rf build
+	echo -e "Build directory has been removed."
+fi
+
 mkdir -p build
 cmake -S . -B build -DBUILD_TEST=OFF
 cmake --build build
-cd build
-cp -ar ./Apps/SizingMain_Windows.exe ../Bin/
-cd ..
+# cd build
+# cp -ar ./Apps/SizingMain_Windows.exe ../Bin/
+# cd ..
